Added output toggling
This introduced a breaking change from the previous beta version.
Metadata was added on top of the hook file which is required. You would
need to re-init a project for this to work now. In the future there will
not be a similar breaking change since the metadata will be append only.